refactor: Jwk 응답 dto 폴더 구조 변경
